Log in improvement and bug fixes
The login page has been improved, with the buttons more neatly organized and spaced better. The footer has also been done. Additionally, the colors have been changed to a light color palette for easier styling. However, a dark mode is going to be planned. Additionally, small issues were fixed. The font family was moved to *, html, body because it removed the redundancy of having to specify the font family in each element. Also, all links that are supposed to have an effect on hovering do have it now.